requirements.txt
aiohttp===3.7.4.post0
anyio===3.5.0
async-timeout===3.0.1
attrs===21.2.0
beautifulsoup4===4.10.0
certifi===2021.10.8
chardet===4.0.0
charset-normalizer===2.0.9
discord.py===1.7.3
h11===0.12.0
httpcore===0.14.7
httpx===0.22.0
idna===3.2
lyricsgenius===3.0.1
multidict===5.2.0
pip===22.0.4
pipenv===2023.4.29
pipreqs===0.4.11
requests===2.26.0
requests-oauthlib===1.3.1
setuptools===67.7.2
sniffio===1.2.0
soupsieve===2.3.1
urllib3===1.26.7
wavelink===0.9.10
wheel===0.37.1
yarl===1.7.0
youtube-search-python===1.6.2
ytmusicapi===0.21.0
